# server/portal/internal/docs/swagger.yaml
basePath: /fe-v1
definitions:
  es.DeviceDTO:
    properties:
      archType:
        type: string
      ciCode:
        type: string
      cluster:
        type: string
      clusterId:
        type: integer
      cpu:
        type: number
      featureCount:
        type: integer
      id:
        type: integer
      ip:
        type: string
      isSpecial:
        type: boolean
      memory:
        type: number
      orderStatus:
        description: 在订单中的状态
        type: string
      role:
        type: string
      status:
        type: string
    type: object
  es.OrderDTO:
    properties:
      actionType:
        description: pool_entry, pool_exit, maintenance_request, maintenance_uncordon
        type: string
      clusterId:
        type: integer
      clusterName:
        type: string
      completionTime:
        type: string
      createdAt:
        type: string
      createdBy:
        type: string
      description:
        description: 订单描述
        type: string
      deviceCount:
        type: integer
      deviceInfo:
        allOf:
        - $ref: '#/definitions/es.DeviceDTO'
        description: DeviceID字段已移除，使用Devices列表和OrderDevice关联表
      devices:
        description: 设备ID列表
        items:
          type: integer
        type: array
      executionTime:
        type: string
      executor:
        type: string
      externalTicketId:
        type: string
      extraInfo:
        additionalProperties: true
        description: 额外信息，用于存储维护原因等
        type: object
      failureReason:
        type: string
      id:
        type: integer
      maintenanceEndTime:
        type: string
      maintenanceStartTime:
        type: string
      name:
        description: 订单名称
        type: string
      orderNumber:
        type: string
      resourcePoolType:
        description: 资源池类型
        type: string
      status:
        type: string
      strategyId:
        type: integer
      strategyName:
        type: string
      strategyThresholdValue:
        type: string
      strategyTriggeredValue:
        type: string
    type: object
...
  order.GeneralOrderCreateDTO:
    properties:
      createdBy:
        type: string
      description:
        type: string
      name:
        type: string
      summary:
        description: Add other fields specific to creating a general order
        type: string
    required:
    - createdBy
    - name
    type: object
  order.GeneralOrderDTO:
    properties:
      completionTime:
        description: 完成时间
        type: string
      createdAt:
        description: 创建时间
        type: string
      createdBy:
        description: 创建人
        type: string
      description:
        description: 订单描述
        type: string
      details:
        $ref: '#/definitions/portal.GeneralOrderDetail'
      elasticScalingDetail:
        allOf:
        - $ref: '#/definitions/portal.ElasticScalingOrderDetail'
        description: 关联关系
      executionTime:
        description: 执行时间
        type: string
      executor:
        description: 执行人
        type: string
      failureReason:
        description: 失败原因
        type: string
      id:
        description: 主键ID
        type: integer
      maintenanceDetail:
        allOf:
        - $ref: '#/definitions/portal.MaintenanceOrderDetail'
        description: 设备维护详情
      name:
        description: 订单名称
        type: string
      orderNumber:
        description: 唯一订单号
        type: string
      status:
        allOf:
        - $ref: '#/definitions/portal.OrderStatus'
        description: 订单状态
      type:
        allOf:
        - $ref: '#/definitions/portal.OrderType'
        description: 订单类型
      updatedAt:
        description: 更新时间
        type: string
    type: object
  order.GeneralOrderQueryDTO:
    properties:
      createdBy:
        type: string
      name:
        description: 订单名称，支持模糊查询
        type: string
      page:
        type: integer
      pageSize:
        type: integer
      status:
        type: string
    type: object
...
  order.UpdateOrderStatusRequest:
    properties:
      reason:
        type: string
      status:
        type: string
    required:
    - status
    type: object
  portal.ElasticScalingOrderDetail:
    properties:
      actionType:
        description: 订单操作类型（入池/退池）
        type: string
      clusterID:
        description: 关联集群ID
        type: integer
      createdAt:
        description: 创建时间
        type: string
      deviceCount:
        description: 请求的设备数量
        type: integer
      id:
        description: 主键ID
        type: integer
      order:
        allOf:
        - $ref: '#/definitions/portal.Order'
        description: 关联关系
      orderID:
        description: 关联订单ID（外键）
        type: integer
      resourcePoolType:
        description: 资源池类型
        type: string
      strategyID:
        description: 关联策略ID（可为NULL）
        type: integer
      strategyThresholdValue:
        description: 策略触发时的阈值设定
        type: string
      strategyTriggeredValue:
        description: 策略触发时的具体指标值
        type: string
      updatedAt:
        description: 更新时间
        type: string
    type: object
  portal.GeneralOrderDetail:
    properties:
      createdAt:
        description: 创建时间
        type: string
      id:
        description: 主键ID
        type: integer
      order:
        allOf:
        - $ref: '#/definitions/portal.Order'
        description: 关联关系
      orderID:
        description: 关联订单ID（外键）
        type: integer
      summary:
        description: 一个简单的摘要字段作为示例
        type: string
      updatedAt:
        description: 更新时间
        type: string
    type: object
  portal.MaintenanceOrderDetail:
    properties:
      clusterID:
        description: 关联集群ID
        type: integer
      comments:
        description: 附加说明
        type: string
      createdAt:
        description: 创建时间
        type: string
      externalTicketID:
        description: 外部工单号
        type: string
      id:
        description: 主键ID
        type: integer
      maintenanceEndTime:
        description: 维护结束时间
        type: string
      maintenanceStartTime:
        description: 维护开始时间
        type: string
      maintenanceType:
        description: 维护类型（cordon/uncordon/general）
        type: string
      order:
        allOf:
        - $ref: '#/definitions/portal.Order'
        description: 关联关系
      orderID:
        description: 关联订单ID（外键）
        type: integer
      priority:
        description: 优先级（high/medium/low）
        type: string
      reason:
        description: 维护原因
        type: string
      updatedAt:
        description: 更新时间
        type: string
    type: object
  portal.Order:
    properties:
      completionTime:
        description: 完成时间
        type: string
      createdAt:
        description: 创建时间
        type: string
      createdBy:
        description: 创建人
        type: string
      description:
        description: 订单描述
        type: string
      elasticScalingDetail:
        allOf:
        - $ref: '#/definitions/portal.ElasticScalingOrderDetail'
        description: 关联关系
      executionTime:
        description: 执行时间
        type: string
      executor:
        description: 执行人
        type: string
      failureReason:
        description: 失败原因
        type: string
      id:
        description: 主键ID
        type: integer
      maintenanceDetail:
        allOf:
        - $ref: '#/definitions/portal.MaintenanceOrderDetail'
        description: 设备维护详情
      name:
        description: 订单名称
        type: string
      orderNumber:
        description: 唯一订单号
        type: string
      status:
        allOf:
        - $ref: '#/definitions/portal.OrderStatus'
        description: 订单状态
      type:
        allOf:
        - $ref: '#/definitions/portal.OrderType'
        description: 订单类型
      updatedAt:
        description: 更新时间
        type: string
    type: object
  portal.OrderStatus:
    enum:
    - pending
    - processing
    - returning
    - return_completed
    - no_return
    - completed
    - failed
    - cancelled
    - ignored
    type: string
    x-enum-comments:
      OrderStatusCancelled: 已取消
      OrderStatusCompleted: 已完成
      OrderStatusFailed: 失败
      OrderStatusIgnored: 已忽略
      OrderStatusNoReturn: 无需归还（退池订单专用）
      OrderStatusPending: 待处理
      OrderStatusProcessing: 处理中
      OrderStatusReturnCompleted: 归还完成（退池订单专用）
      OrderStatusReturning: 归还中（退池订单专用）
    x-enum-varnames:
    - OrderStatusPending
    - OrderStatusProcessing
    - OrderStatusReturning
    - OrderStatusReturnCompleted
    - OrderStatusNoReturn
    - OrderStatusCompleted
    - OrderStatusFailed
    - OrderStatusCancelled
    - OrderStatusIgnored
  portal.OrderType:
    enum:
    - elastic_scaling
    - maintenance
    - deployment
    - general
    type: string
    x-enum-comments:
      OrderTypeDeployment: 应用部署
      OrderTypeElasticScaling: 弹性伸缩
      OrderTypeGeneral: 通用订单
      OrderTypeMaintenance: 设备维护
    x-enum-varnames:
    - OrderTypeElasticScaling
    - OrderTypeMaintenance
    - OrderTypeDeployment
    - OrderTypeGeneral
  render.ErrorResponse:
    properties:
      error:
        example: 操作失败
        type: string
    type: object
  render.Response:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Navy-NG 管理平台 API 文档
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Navy-NG API
  version: "1.0"
paths:
...
  /fe-v1/elastic-scaling/orders:
    get:
      consumes:
      - application/json
      description: 获取弹性伸缩订单列表，支持分页和过滤
      parameters:
      - description: 页码，默认1
        in: query
        name: page
        type: integer
      - description: 每页大小，默认10
        in: query
        name: pageSize
        type: integer
      - description: 集群ID
        in: query
        name: clusterId
        type: integer
      - description: 策略ID
        in: query
        name: strategyId
        type: integer
      - description: 动作类型
        in: query
        name: actionType
        type: string
      - description: 订单状态
        in: query
        name: status
        type: string
      - description: 订单名称（支持模糊搜索）
        in: query
        name: name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/render.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/render.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/render.ErrorResponse'
      summary: 获取弹性伸缩订单列表
      tags:
      - 弹性伸缩订单
    post:
      consumes:
      - application/json
      description: 创建新的弹性伸缩订单
      parameters:
      - description: 订单数据
        in: body
        name: order
        required: true
        schema:
          $ref: '#/definitions/es.OrderDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/render.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/render.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/render.ErrorResponse'
      summary: 创建弹性伸缩订单
      tags:
      - 弹性伸缩订单
  /fe-v1/elastic-scaling/orders/{id}:
    get:
      consumes:
      - application/json
      description: 获取指定订单的详细信息
      parameters:
      - description: 订单ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/render.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/render.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/render.ErrorResponse'
      summary: 获取弹性伸缩订单详情
      tags:
      - 弹性伸缩订单
  /fe-v1/elastic-scaling/orders/{id}/devices:
    get:
      consumes:
      - application/json
      description: 获取指定订单关联的设备列表
      parameters:
      - description: 订单ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/render.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/render.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/render.ErrorResponse'
      summary: 获取订单关联的设备
      tags:
      - 弹性伸缩订单
  /fe-v1/elastic-scaling/orders/{id}/devices/{device_id}/status:
    put:
      consumes:
      - application/json
      description: 更新指定订单中指定设备的状态
      parameters:
      - description: 订单ID
        in: path
        name: id
        required: true
        type: integer
      - description: 设备ID
        in: path
        name: device_id
        required: true
        type: integer
      - description: 设备状态更新请求
        in: body
        name: request
        required: true
        schema:
          properties:
            status:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/render.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/render.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/render.ErrorResponse'
      summary: 更新订单关联设备的状态
      tags:
      - 弹性伸缩订单
  /fe-v1/elastic-scaling/orders/{id}/status:
    put:
      consumes:
      - application/json
      description: 更新指定订单的状态
      parameters:
      - description: 订单ID
        in: path
        name: id
        required: true
        type: integer
      - description: 状态更新请求
        in: body
        name: request
        required: true
        schema:
          properties:
            reason:
              type: string
            status:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/render.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/render.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/render.ErrorResponse'
      summary: 更新订单状态
      tags:
      - 弹性伸缩订单
...
  /fe-v1/elastic-scaling/stats/orders:
    get:
      consumes:
      - application/json
      description: 获取不同时间范围的订单统计数据
      parameters:
      - description: 时间范围（7d/30d/90d）
        in: query
        name: timeRange
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/render.Response'
      summary: 获取订单统计
      tags:
      - 弹性伸缩
...
  /fe-v1/orders/{orderType}:
    get:
      consumes:
      - application/json
      description: 根据订单类型和查询参数获取订单列表
      parameters:
      - description: 订单类型 (e.g., general)
        in: path
        name: orderType
        required: true
        type: string
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页数量
        in: query
        name: pageSize
        type: integer
      - description: 订单状态
        in: query
        name: status
        type: string
      - description: 创建者
        in: query
        name: createdBy
        type: string
      - description: 订单名称，支持模糊查询
        in: query
        name: name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功时返回订单列表和总数
          schema:
            allOf:
            - $ref: '#/definitions/render.Response'
            - properties:
                data:
                  $ref: '#/definitions/order.GeneralOrderQueryDTO'
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/render.ErrorResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/render.ErrorResponse'
      summary: 获取指定类型的订单列表
      tags:
      - 统一订单
    post:
      consumes:
      - application/json
      description: 根据指定的订单类型创建一个新的订单
      parameters:
      - description: 订单类型 (e.g., general)
        in: path
        name: orderType
        required: true
        type: string
      - description: 创建通用订单所需的数据
        in: body
        name: order
        required: true
        schema:
          $ref: '#/definitions/order.GeneralOrderCreateDTO'
      produces:
      - application/json
      responses:
        "201":
          description: 成功时返回创建的订单详情
          schema:
            allOf:
            - $ref: '#/definitions/render.Response'
            - properties:
                data:
                  $ref: '#/definitions/order.GeneralOrderDTO'
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/render.ErrorResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/render.ErrorResponse'
      summary: 创建一个新订单
      tags:
      - 统一订单
  /fe-v1/orders/{orderType}/{id}:
    get:
      consumes:
      - application/json
      description: 根据订单类型和ID获取订单的详细信息
      parameters:
      - description: 订单类型 (e.g., general)
        in: path
        name: orderType
        required: true
        type: string
      - description: 订单ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功时返回通用订单详情
          schema:
            allOf:
            - $ref: '#/definitions/render.Response'
            - properties:
                data:
                  $ref: '#/definitions/order.GeneralOrderDTO'
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/render.ErrorResponse'
        "404":
          description: 订单不存在
          schema:
            $ref: '#/definitions/render.ErrorResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/render.ErrorResponse'
      summary: 获取特定类型的订单详情
      tags:
      - 统一订单
...
  /fe-v1/orders/{orderType}/{id}/status:
    put:
      consumes:
      - application/json
      description: 更新指定订单的状态
      parameters:
      - description: 订单类型 (e.g., general)
        in: path
        name: orderType
        required: true
        type: string
      - description: 订单ID
        in: path
        name: id
        required: true
        type: integer
      - description: 状态更新请求
        in: body
        name: statusUpdate
        required: true
        schema:
          $ref: '#/definitions/order.UpdateOrderStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/render.Response'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/render.ErrorResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/render.ErrorResponse'
      summary: 更新订单状态
      tags:
      - 统一订单
...
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"